Create .gitignore for new main
As a result of following outdated tutorials, I was pushing to the wrong branch: instead of Main, I was pushing to a certain Master that the git tutorial came up with. 
So after renaming the Master repository to Main and deleting the old main repository, I realized that I forgot to add the .gitignore file (which I had made for the old Main repository).
I am now committing this new .gitignore to the updated Main repository.